feat: updated in modal with the new tabs, buy and sell.
