+++ package.json
@@ -42,13 +42,14 @@
   "devDependencies": {
     "@types/node": "^17.0.10",
     "@types/react": "^17.0.38",
-    "eslint": "8.7.0",
+    "eslint": "^8.7.0",
     "eslint-config-next": "12.0.8",
     "eslint-config-prettier": "^8.3.0",
     "husky": "^7.0.4",
     "lint-staged": "^12.3.2",
     "miragejs": "^0.1.43",
-    "prettier": "^2.5.1",
+    "prettier": "2.6.2",
+    "tsconfig-paths": "^3.12.0",
     "typescript": "^4.5.4"
   },
   "husky": {
