+++ src/Modules/Home/components/ModalExchange/components/BuyCryptocurrenciesTab/index.ts
@@ -1,0 +1,3 @@
+import { BuyCryptocurrenciesTab } from './BuyCryptocurrenciesTab';
+
+export default BuyCryptocurrenciesTab;
