+++ src/Modules/Home/components/ModalExchange/components/BuyCryptocurrenciesTab/styles.module.scss
@@ -47,6 +47,11 @@
     }
 }
 
+.totalWithSelect {
+    color: #a7a5a5;
+    background:#e6e6e6 !important;
+}
+
 .selectCoin{
     margin-bottom: 2rem;
     background-color: #dddddd;
@@ -75,7 +80,7 @@
     align-items: center;
     justify-content: center;
 
-    margin-top: 20px;
+    margin-top: 2rem;
 
     button {
         padding: 8px 30px;
@@ -119,23 +124,4 @@
     border-radius: 0.5rem;
     padding: 0.6rem;
     background-color: rgb(96, 194, 57);
-}
-
-.seeGraph {
-    color: #f1f0f0;
-    font-weight: bold;
-    cursor: pointer;
-    padding: 10px; 
-    border-radius: 5px; 
-    background-color: rgb(87, 75, 253);
-    margin: auto;
-    text-align: center;
-    font-family: 'Courier New', Courier, monospace;
-    width: 180px;
-
-    transition: filter 0.3s;
-
-    &:hover {
-        filter: brightness(0.9);
-    }
 }
