+++ src/Modules/Home/components/ModalExchange/components/Graph/index.ts
@@ -1,0 +1,3 @@
+import { Graph} from './Graph';
+
+export default Graph;
