+++ src/Modules/Home/components/ModalExchange/components/SellCryptocurrenciesTab/index.ts
@@ -1,0 +1,3 @@
+import { SellCryptocurrenciesTab } from './SellCryptocurrenciesTab';
+
+export default SellCryptocurrenciesTab;
