+++ src/Modules/Home/components/ModalExchange/components/SellCryptocurrenciesTab/styles.module.scss
@@ -120,22 +120,3 @@
     padding: 0.6rem;
     background-color: rgb(96, 194, 57);
 }
-
-.seeGraph {
-    color: #f1f0f0;
-    font-weight: bold;
-    cursor: pointer;
-    padding: 10px; 
-    border-radius: 5px; 
-    background-color: rgb(87, 75, 253);
-    margin: auto;
-    text-align: center;
-    font-family: 'Courier New', Courier, monospace;
-    width: 180px;
-
-    transition: filter 0.3s;
-
-    &:hover {
-        filter: brightness(0.9);
-    }
-}
