+++ src/Modules/Transactions/TransactionsPage.tsx
@@ -8,8 +8,7 @@ import Container from './components/Container';
  * @name TransactionsPage
  * 
  * @description
- * Esse componente irá carregar os dados de todas
- * as transações realizadas na HomePage.
+ * Component responsible for rendering the transactions page.
  */
 export const TransactionsPage = (): JSX.Element => {
     return (
