+++ src/Modules/Transactions/components/Body/Body.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useTransactions } from '../../../Provider/useTransactions';
+import { useTransactions } from '@Provider/useTransactions';
 
 /**
  * @export
@@ -7,8 +7,8 @@ import { useTransactions } from '../../../Provider/useTransactions';
  * @name Body
  * 
  * @description
- * Componente responsável por montar o Body da
- * tela de transações.
+ * Component responsible for mounting the body of
+ * transactions page.
  */
 export const Body = () => {
     const { values } = useTransactions();
