+++ tsconfig.json
@@ -13,7 +13,16 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "jsx": "preserve",
-    "incremental": true
+    "incremental": true,
+    "baseUrl": ".",
+    "paths": {
+      "@Modules/*": ["./src/Modules/*"],
+      "@components/*": ["./src/components/*"],
+      "@Home/*": ["./src/Modules/Home/*"],
+      "@Transactions/*": ["./src/Modules/Transactions/*"],
+      "@Provider/*": ["./src/Modules/Provider/*"],
+      "@services/*": ["./src/services/*"],
+    },
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
   "exclude": ["node_modules"]
